Don't try to write a GIF file as a M4A cover art
taglib-ruby does not support writing GIF files to a TagLib::MP4::Item
so do not try doing it.

A better solution would be to do what iTunes does, that is, convert a
GIF to a PNG before storing it as cover art of an M4A. This might
happen in a future commit. For now, don't store the image in the M4A